refactor(profile): consolidate profile page functionality and improve loading state
- Deleted the redundant `page copy.tsx` file to streamline the codebase.
- Updated the `page.tsx` component to enhance loading state handling and ensure proper user feedback during session retrieval.
- Improved session management by simplifying the component structure and ensuring unauthenticated users receive appropriate messages.
- Enhanced tab rendering logic for better user experience and maintainability.
